pyproject.toml
[project]
name = "credit-investment-memo-agent"
version = "0.1.0"
description = "AI-powered credit investment memo agent built with LangChain Deep Agents"
requires-python = ">=3.11"
dependencies = [
    "deepagents",
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langsmith>=0.3.0",
    "langchain-openai",
    "langchain-perplexity",
    "langchain-tavily",
    "langchain-text-splitters",
    "langchain-chroma",
    "python-docx",
    "python-dotenv",
]
